Added a temporary check for now to test if the theory is sound.

+++ providers/node.rb
@@ -15,10 +15,17 @@ action :create do
   lxc_base     = node[:lxc_container][:path] + "/"+new_resource.lxc_name
   rootfs       = lxc_base + "/rootfs"
   lxc_conf     = lxc_base + "/config"
-  lxc_opts = ""
+  lxc_opts     = ""
+  is_setup     = false
 
   if (new_resource.lxc_ver)
     lxc_opts = "-- --release=#{new_resource.lxc_ver}"
+  end
+
+
+  execute "check-exists-#{new_resource.lxc_name}" do
+    check = "lxc-ls | grep #{new_resource.lxc_name}"
+    is_setup = true if check
   end
 
 
@@ -36,6 +43,7 @@ action :create do
   execute "backup-#{lxc_conf}" do
     command "mv #{lxc_conf} #{lxc_conf}.dist"
     not_if { ::File.exists?("#{lxc_conf}.dist") }
+    not_if is_setup == true
   end
 
 
@@ -48,12 +56,25 @@ action :create do
   # can generate via: openssl rand -hex 6 | sed 's/\(..\)/\1:/g; s/.$//'
   ruby_block "mac_addr_#{new_resource.name}" do
     block do
-      read_mac = ::File.readlines("#{lxc_conf}.dist").grep(/^lxc.network.hwaddr/)
-      put_mac  = read_mac[0].split(/=/).join("").sub(/.*?\s+/, "").chomp
+      #read_mac = ::File.readlines("#{lxc_conf}.dist").grep(/^lxc.network.hwaddr/)
+      #put_mac  = read_mac[0].split(/=/).join("").sub(/.*?\s+/, "").chomp
+      def generate_mac
+        mac = `openssl rand -hex 6`.chomp
+        mac = mac.scan(/.{1,2}/).join(":")
+
+        if (mac =~ /^fe:/)
+          return mac
+        else
+          generate_mac
+        end
+      end
+
+      gen_mac = generate_mac
 
       # Reset MAC to the LXC generated one:
-      node.set[:lxc_container][:node][:"#{new_resource.lxc_name}"][:hwaddr] = put_mac
+      node.set[:lxc_container][:node][:"#{new_resource.lxc_name}"][:hwaddr] = gen_mac
     end
+    not_if is_setup == true
   end
 
 
@@ -65,6 +86,7 @@ action :create do
     block do
       mac_addr = node[:lxc_container][:node][:"#{new_resource.lxc_name}"][:hwaddr]
     end
+    not_if is_setup == true
   end
 
 
@@ -80,12 +102,14 @@ action :create do
     })
 
     only_if { ::File.exists?("#{lxc_conf}.dist") }
+    not_if is_setup == true
   end
 
 
   execute "launch-lxc-#{new_resource.lxc_name}" do
     command "lxc-start -n #{new_resource.lxc_name} -d"
     not_if "lxc-ls --active | grep #{new_resource.lxc_name}"
+    not_if is_setup == true
   end
 end
 
@@ -94,5 +118,6 @@ action :destroy do
   execute "destroy-lxc-#{new_resource.lxc_name}" do
     command "lxc-stop -k -n #{new_resource.lxc_name}; lxc-destroy -n #{new_resource.lxc_name}"
     only_if "lxc-ls | grep #{new_resource.lxc_name}"
+    only_if is_setup == true
   end
 end
